pyproject.toml
[project]
name = "tts-ou"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "google-cloud-texttospeech>=2.33.0",
    "google-generativeai>=0.8.5",
    "markdown>=3.10",
    "markitdown>=0.0.2",
    "python-dotenv>=1.2.1",
    "typer>=0.20.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.14.9",
]
